refactor: Update import paths and enhance modularity across scripts
- Refactored import statements to utilize new shared module paths, improving code organization and maintainability.
- Introduced new shared modules for core functionalities, including logger, path helpers, and validation utilities, ensuring consistent usage throughout the codebase.
- Updated various scripts to reflect these changes, enhancing clarity and reducing complexity in the overall structure.
- Improved error handling and logging mechanisms in the validation and PDF generation processes for better feedback during operations.
