+++ scripts/shared/core/logger.ts
