+++ scripts/shared/core/path-helpers.ts
