+++ scripts/shared/data/context-operations.ts
@@ -1,12 +1,12 @@
 import { writeFileSync } from 'fs';
 import yaml from 'js-yaml';
 import type { z } from 'zod';
-import type { Result, FileToValidateWithYamlData, SetContextSuccess } from './validation-pipeline';
-import { formatZodError } from './validation-pipeline';
-import { tryCatch } from './functional-utils';
-import { MetadataSchema } from '../../src/zod/schemas';
-import { TailorContextSchema, type TailorContext } from '../../src/zod/tailor-context-schema';
-import type { CompanyFileValue } from './config';
+import type { Result, FileToValidateWithYamlData, SetContextSuccess } from '../validation/validation-pipeline';
+import { formatZodError } from '../validation/validation-pipeline';
+import { tryCatch } from '../core/functional-utils';
+import { MetadataSchema } from '../../../src/zod/schemas';
+import { TailorContextSchema, type TailorContext } from '../../../src/zod/tailor-context-schema';
+import type { CompanyFileValue } from '../core/config';
 
 /**
  * Generates YAML content for tailor-context.yaml using schema-driven approach.
