+++ scripts/shared/data/data-generation.ts
@@ -1,14 +1,14 @@
 import { pipe } from 'remeda';
-import type { Result, FileToValidateWithYamlData } from './validation-pipeline';
-import { chain, chainPipe, tryCatch } from './functional-utils';
-import { validateApplicationData } from '../../src/zod/validation';
+import type { Result, FileToValidateWithYamlData } from '../validation/validation-pipeline';
+import { chain, chainPipe, tryCatch } from '../core/functional-utils';
+import { validateApplicationData } from '../../../src/zod/validation';
 import {
   transformFilesToApplicationData,
   generateTypeScriptModule,
-} from '../utils/application-data-generator';
-import type { ApplicationData } from '../../src/types';
-import { loggers } from './logger';
-import { PATHS } from './config';
+} from './application-data-generator';
+import type { ApplicationData } from '../../../src/types';
+import { loggers } from '../core/logger';
+import { PATHS } from '../core/config';
 
 /**
  * Transforms validated YAML files to ApplicationData using fileName mapping.
