+++ scripts/shared/handlers/result-handlers.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod';
 import { match } from 'ts-pattern';
-import type { Result } from './validation-pipeline';
-import type { SetContextSuccess } from './validation-pipeline';
-import type { ValidationOnlySuccess } from './validation-only-pipeline';
-import type { Logger } from './logger';
+import type { Result } from '../validation/validation-pipeline';
+import type { SetContextSuccess } from '../validation/validation-pipeline';
+import type { ValidationOnlySuccess } from '../validation/validation-only-pipeline';
+import type { Logger } from '../core/logger';
 
 /**
  * Options for configuring pipeline error handling behavior.
